requirements.txt
discord.py>=2.5.0
motor>=3.7.0
jishaku>=2.6.0